添加用户分组;修改Clue中显示任务的方式

+++ database/migrations/2021_11_02_161424_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('title')->nullable();
-            $table->integer('price')->nullable();
+            $table->decimal('price')->nullable();
             $table->boolean('show')->default(1);
             $table->integer('sales')->nullable();
